+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Store IDs of payout transactions in database.
 - Update subscription only when payout transaction is confirmed.
 - Re-open closed invoices when address receives new payment.
+- Add `get-payment-address` command.
 
 ### Changed